SW_lab/sw_lab_part4/servizio_telegram/requirements.txt
# Required
cherrypy == 18.6.0
paho-mqtt == 1.5.0
prompt-toolkit == 3.0.6
python-telegram-bot == 12.8
requests == 2.24.0

